split up logic into files
